backend/project_name/requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.5.0
APScheduler==3.10.4
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.4.2
boto3==1.34.96
botocore==1.34.96
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.8
ecdsa==0.19.0
fastapi==0.110.3
filelock==3.14.0
frozendict==2.4.4
google-api-core==2.19.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
identify==2.5.36
idna==3.7
jmespath==1.0.1
kombu==5.3.7
lxml==5.2.2
Mako==1.3.3
MarkupSafe==2.1.5
multitasking==0.0.11
mypy==1.10.0
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.2
passlib==1.7.4
pathspec==0.12.1
peewee==3.17.5
pipenv==2023.12.1
platformdirs==4.2.1
polygon-api-client==1.13.7
pre-commit==3.7.1
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycurl==7.45.3
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.7
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
typing==3.7.4.3
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.29.0
vine==5.1.0
virtualenv==20.26.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
yfinance==0.2.40
requests-cache==1.2.0
requests-ratelimiter==0.6.0
pyrate-limiter==2.10.0
